AxmolMediaPlayer is shareable
